Add splash screen and navigation setup
