+++ pubspec.yaml
@@ -17,3 +17,6 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+  assets:
+    - ./assets/images/petut.png
+
